Fazendo atualizações em cascata

+++ algaworks-ecommerce/src/main/java/com/algaworks/ecommerce/model/ItemPedido.java
@@ -6,7 +6,6 @@ import lombok.Setter;
 
 import javax.persistence.*;
 import java.math.BigDecimal;
-import java.math.BigInteger;
 
 @Getter
 @Setter
@@ -32,6 +31,6 @@ public class ItemPedido {
     private BigDecimal precoProduto;
 
     @Column(nullable = false) // integer not null
-    private BigInteger quantidade;
+    private Integer quantidade;
 
 }
